Add target_health_protocol to ALB
Sometimes we want to change the protocol of target_health_protocol to something else than HTTP (eg HTTPS). This change makes it possible. In 99% of the cases the target_health_protocol will be the same as the target_protocol. If this is the case, by working with a condition you only need to set the target_protocol. You can still override it when you need it.